Introduce "final" middleware
Express doesn't require that the user specifies
`next` all the time.
I think I was concerned that the compiler wouldn't
allow the overload and end up in ambiguities, but
that doesn't seem to be a problem (anymore), so
let's add middleware that only takes `(req, res)`.

(Is that still middleware? finalware maybe :-) ).